+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Internal
 
+- Added cypress text for Expiration Date @iFlameing
+
 ## 6.0.0 (2020-05-18)
 
 ### Breaking